Fix almost all -Wall warnings

+++ Externals/ode/ode/src/timer.cpp
@@ -407,7 +407,7 @@ void dTimerReport (FILE *fout, int average)
       t = total;
       p = 100.0;
     }
-    fprintf (fout,"%-*s %7.2fms %6.2f%%",maxl,event[i].description,
+    fprintf (fout,"%-*s %7.2fms %6.2f%%", static_cast<int>(maxl), event[i].description,
 	     t*ccunit * 1000.0, p);
     if (average && i < (num-1)) {
       fprintf (fout,"  (avg %7.2fms %6.2f%%)",
