+++ src/xrGame/mp_config_sections.cpp
@@ -25,7 +25,7 @@ mp_config_sections::mp_config_sections() : m_tmp_dumper(NULL, FALSE, FALSE, FALS
             m_mp_sections.push_back(_GetItem(line, j, tmp_single_item));
         }
     }
-    for (int i = 0; i < (sizeof(important_sections) / sizeof(char*)); ++i)
+    for (int i = 0; i < static_cast<int>(sizeof(important_sections) / sizeof(char*)); ++i)
     {
         m_mp_sections.push_back(shared_str(important_sections[i]));
     }
